+++ custom_components/ember_mug/coordinator.py
@@ -1,21 +1,26 @@
 """Coordinator for all the sensors."""
 from __future__ import annotations
 
-from datetime import timedelta
+import asyncio
 import logging
-from typing import Any
+from datetime import timedelta
+from typing import TYPE_CHECKING, Any
 
+from bleak import BleakError
 from bleak_retry_connector import close_stale_connections
-from ember_mug import EmberMug
 from ember_mug.data import MugData
-from home_assistant_bluetooth import BluetoothServiceInfoBleak
-from homeassistant.components.bluetooth import BluetoothChange
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.device_registry import CONNECTION_BLUETOOTH
 from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import MANUFACTURER
+
+if TYPE_CHECKING:
+    from ember_mug import EmberMug
+    from home_assistant_bluetooth import BluetoothServiceInfoBleak
+    from homeassistant.components.bluetooth import BluetoothChange
+
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -68,6 +73,14 @@ class MugDataUpdateCoordinator(DataUpdateCoordinator[MugData]):
                 changed += await self.mug.update_queued_attributes()
             self._last_refresh_was_full = not self._last_refresh_was_full
             self.available = True
+        except (asyncio.TimeoutError, BleakError) as e:
+            if self.available is True:
+                _LOGGER.debug("%s is not available: %s", e)
+                self.available = False
+            if self._initial_update is True:
+                raise UpdateFailed(
+                    f"An error occurred updating {self.mug.model_name}: {e=}",
+                ) from e
         except Exception as e:
             _LOGGER.error(
                 "An error occurred whilst updating the %s: %s",
@@ -77,7 +90,7 @@ class MugDataUpdateCoordinator(DataUpdateCoordinator[MugData]):
             self.available = False
             raise UpdateFailed(
                 f"An error occurred updating {self.mug.model_name}: {e=}",
-            )
+            ) from e
 
         _LOGGER.debug(
             "[%s Update] Changed: %s",
@@ -92,7 +105,7 @@ class MugDataUpdateCoordinator(DataUpdateCoordinator[MugData]):
         service_info: BluetoothServiceInfoBleak,
     ) -> None:
         """Handle the device going unavailable."""
-        _LOGGER.warning("%s is unavailable", self.mug.model_name)
+        _LOGGER.debug("%s is unavailable", self.mug.model_name)
         self.available = False
         self.async_update_listeners()
 
@@ -137,9 +150,7 @@ class MugDataUpdateCoordinator(DataUpdateCoordinator[MugData]):
         firmware = self.data.firmware
         return DeviceInfo(
             connections={(CONNECTION_BLUETOOTH, self.mug.device.address)},
-            name=name
-            if (name := self.data.name) and name != "EMBER"
-            else self.device_name,
+            name=name if (name := self.data.name) and name != "EMBER" else self.device_name,
             model=self.data.model.name,
             suggested_area="Kitchen",
             hw_version=str(firmware.hardware) if firmware else None,